Move helper methods inside the main app file
